backend_main/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asyncio==4.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-multiselectfield==1.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.4
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
python-openid==2.2.5
python-ulid==1.1.0
redis==5.3.1
redis-om==0.3.5
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
types-cffi==1.17.0.20250822
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
zope.interface==8.0.1
